Minor refactor to the referral overview component
