+++ client/a8c-for-agencies/sections/referrals/common/step-section-item/style.scss
@@ -73,23 +73,23 @@
 		margin-left: auto;
 	}
 
-	.badge {
+	&.badge {
 		white-space: nowrap;
 		position: relative;
 		top: -1px;
 	}
 
-	.badge--success {
+	&.badge--success {
 		background-color: var(--color-success-5);
 		color: var(--color-success-80);
 	}
 
-	.badge--warning {
+	&.badge--warning {
 		background-color: var(--color-warning-5);
 		color: var(--color-warning-80);
 	}
 
-	.badge--error {
+	&.badge--error {
 		background-color: var(--color-error-5);
 		color: var(--color-error-80);
 	}
