+++ client/a8c-for-agencies/sections/referrals/lib/constants.ts
@@ -1,0 +1,3 @@
+import { A4A_MARKETPLACE_PRODUCTS_LINK } from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
+
+export const REFER_PRODUCTS_LINK = `${ A4A_MARKETPLACE_PRODUCTS_LINK }?purchase-type=referral`;
